cognitus_flask/requirements.txt
amqp==2.6.1
billiard==3.6.3.0
celery==4.4.7
click==7.1.2
Flask==1.1.2
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
kombu==4.6.11
MarkupSafe==1.1.1
numpy==1.19.1
psycopg2-binary==2.8.5
pytz==2020.1
redis==3.5.3
scikit-learn==0.23.2
scipy==1.5.2
threadpoolctl==2.1.0
vine==1.3.0
Werkzeug==1.0.1
zipp==3.1.0
